# apps/api/generated/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Qafiyah API",
    "description": "Public, read-only REST API for the Qafiyah Arabic classical-poetry catalog.\n\nBrowse poems and poets and filter them by the taxonomies they belong to: eras, meters, rhymes, themes, and collections. Full-text search spans both poems and poets.\n\nConventions:\n- All endpoints are GET and require no authentication.\n- Resources are addressed by stable `slug` identifiers returned in list responses.\n- List endpoints are paginated via a 1-based `page` query param and return a `pagination` block.\n- Multi-select filters are repeatable array params, e.g. `?era=abbasi&era=andalusi`.\n- Errors follow RFC 9457 problem details and are served as `application/problem+json`.\n- Requests are rate limited. Rate-limited responses carry `x-ratelimit-limit`, `x-ratelimit-remaining`, and `x-ratelimit-reset`.",
    "contact": {
      "name": "Qafiyah",
      "url": "https://qafiyah.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/v1"
    }
  ],
  "paths": {
    "/collections": {
      "get": {
        "tags": [
          "collections"
        ],
        "description": "All curated poem collections (al-dawawin) with poem counts.",
        "operationId": "collections.list",
        "responses": {
          "200": {
            "description": "All collections.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListEnvelope_PoemCountStats"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "headers": {
              "retry-after": {
                "schema": {
                  "type": "string"
                },
                "description": "Seconds until the window resets"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "string"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "string"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "string"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/collections/{slug}": {
      "get": {
        "tags": [
          "collections"
        ],
        "description": "A single collection with poem count, by slug.",
        "operationId": "collections.get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Resource identifier taken from the `slug` field of the matching list endpoint.",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z][a-z-]*$"
            },
            "example": "almuallaqat"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested collection.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemEnvelope_PoemCountStats"
                }
              }
            }
          },
          "400": {
            "description": "Input validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "headers": {
              "retry-after": {
                "schema": {
                  "type": "string"
                },
                "description": "Seconds until the window resets"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "string"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "string"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "string"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/eras": {
      "get": {
        "tags": [
          "eras"
        ],
        "description": "All literary eras (al-usur al-adabiyya) with poem and poet counts.",
        "operationId": "eras.list",
        "responses": {
          "200": {
            "description": "All eras.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListEnvelope_CountedStats"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "headers": {
              "retry-after": {
                "schema": {
                  "type": "string"
                },
                "description": "Seconds until the window resets"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "string"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "string"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "string"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/eras/{slug}": {
      "get": {
        "tags": [
          "eras"
        ],
        "description": "A single era with poem and poet counts, by slug.",
        "operationId": "eras.get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Resource identifier taken from the `slug` field of the matching list endpoint.",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z][a-z-]*$"
            },
            "example": "abbasi"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested era.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemEnvelope_CountedStats"
                }
              }
            }
          },
          "400": {
            "description": "Input validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "headers": {
              "retry-after": {
                "schema": {
                  "type": "string"
                },
                "description": "Seconds until the window resets"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "string"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "string"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "string"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/meters": {
      "get": {
        "tags": [
          "meters"
        ],
        "description": "All prosodic meters (al-buhur) with poem and poet counts.",
        "operationId": "meters.list",
        "responses": {
          "200": {
            "description": "All meters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListEnvelope_CountedStats"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "headers": {
              "retry-after": {
                "schema": {
                  "type": "string"
                },
                "description": "Seconds until the window resets"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "string"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "string"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "string"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/meters/{slug}": {
      "get": {
        "tags": [
          "meters"
        ],
        "description": "A single meter with poem and poet counts, by slug.",
        "operationId": "meters.get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Resource identifier taken from the `slug` field of the matching list endpoint.",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z][a-z-]*$"
            },
            "example": "altawil"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested meter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemEnvelope_CountedStats"
                }
              }
            }
          },
          "400": {
            "description": "Input validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "headers": {
              "retry-after": {
                "schema": {
                  "type": "string"
                },
                "description": "Seconds until the window resets"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "string"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "string"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "string"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/rhymes": {
      "get": {
        "tags": [
          "rhymes"
        ],
        "description": "All rhyme letters (al-qawafi) with poem and poet counts.",
        "operationId": "rhymes.list",
        "responses": {
          "200": {
            "description": "All rhymes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListEnvelope_CountedStats"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "headers": {
              "retry-after": {
                "schema": {
                  "type": "string"
                },
                "description": "Seconds until the window resets"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "string"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "string"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "string"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/rhymes/{slug}": {
      "get": {
        "tags": [
          "rhymes"
        ],
        "description": "A single rhyme with poem and poet counts, by slug.",
        "operationId": "rhymes.get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Resource identifier taken from the `slug` field of the matching list endpoint.",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z][a-z-]*$"
            },
            "example": "meem"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested rhyme.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemEnvelope_CountedStats"
                }
              }
            }
          },
          "400": {
            "description": "Input validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "headers": {
              "retry-after": {
                "schema": {
                  "type": "string"
                },
                "description": "Seconds until the window resets"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "string"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "string"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "string"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/themes": {
      "get": {
        "tags": [
          "themes"
        ],
        "description": "All thematic categories (al-aghrad) with poem counts.",
        "operationId": "themes.list",
        "responses": {
          "200": {
            "description": "All themes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListEnvelope_PoemCountStats"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "headers": {
              "retry-after": {
                "schema": {
                  "type": "string"
                },
                "description": "Seconds until the window resets"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "string"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "string"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "string"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/themes/{slug}": {
      "get": {
        "tags": [
          "themes"
        ],
        "description": "A single theme with poem count, by slug.",
        "operationId": "themes.get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Resource identifier taken from the `slug` field of the matching list endpoint.",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z][a-z-]*$"
            },
            "example": "alnasib"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested theme.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemEnvelope_PoemCountStats"
                }
              }
            }
          },
          "400": {
            "description": "Input validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "headers": {
              "retry-after": {
                "schema": {
                  "type": "string"
                },
                "description": "Seconds until the window resets"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "string"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "string"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "string"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ItemEnvelope_PoemCountStats": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "name",
              "slug",
              "poemsCount"
            ],
            "properties": {
              "name": {
                "type": "string"
              },
              "poemsCount": {
                "type": "integer",
                "format": "int32",
                "example": 47457
              },
              "slug": {
                "type": "string",
                "example": "alnasib",
                "pattern": "^[a-z][a-z-]*$"
              }
            }
          }
        }
      },
...
      "ListEnvelope_PoemCountStats": {
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "name",
                "slug",
                "poemsCount"
              ],
              "properties": {
                "name": {
                  "type": "string"
                },
                "poemsCount": {
                  "type": "integer",
                  "format": "int32",
                  "example": 47457
                },
                "slug": {
                  "type": "string",
                  "example": "alnasib",
                  "pattern": "^[a-z][a-z-]*$"
                }
              }
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
...
    }
  },
...
}