# apps/api/src/routes/taxonomy.rs
use axum::Json;
use axum::extract::{Extension, State};

use crate::domain::taxonomy::{self, Counted, CountedStats, PoemCountStats, PoemCounted};
use crate::envelope::{ItemEnvelope, ListEnvelope, build_pagination};
use crate::error::AppError;
use crate::extract::SafePath;
use crate::log::LogHandle;
use crate::openapi::{ListErrors, LookupErrors};
use crate::slug;
use crate::state::AppState;

fn listed<T>(rows: Vec<T>, log: &LogHandle) -> Json<ListEnvelope<T>> {
    let count = u32::try_from(rows.len()).unwrap_or(u32::MAX);
    log.set("result_count", count);
    Json(ListEnvelope {
        data: rows,
        pagination: build_pagination(1, count.max(1), count),
    })
}

async fn list_counted_kind(
    State(state): State<AppState>,
    Extension(log): Extension<LogHandle>,
    kind: Counted,
) -> Result<Json<ListEnvelope<CountedStats>>, AppError> {
    Ok(listed(taxonomy::list_counted(&state.pg, kind).await?, &log))
}

async fn get_counted_kind(
    State(state): State<AppState>,
    Extension(log): Extension<LogHandle>,
    SafePath(raw): SafePath<String>,
    kind: Counted,
) -> Result<Json<ItemEnvelope<CountedStats>>, AppError> {
    let slug = slug::transliterated(&raw)?;
    log.set(kind.log_field(), slug);
    Ok(Json(ItemEnvelope {
        data: taxonomy::get_counted(&state.pg, kind, slug).await?,
    }))
}

async fn list_poem_counted_kind(
    State(state): State<AppState>,
    Extension(log): Extension<LogHandle>,
    kind: PoemCounted,
) -> Result<Json<ListEnvelope<PoemCountStats>>, AppError> {
    Ok(listed(
        taxonomy::list_by_poem_count(&state.pg, kind).await?,
        &log,
    ))
}

async fn get_poem_counted_kind(
    State(state): State<AppState>,
    Extension(log): Extension<LogHandle>,
    SafePath(raw): SafePath<String>,
    kind: PoemCounted,
) -> Result<Json<ItemEnvelope<PoemCountStats>>, AppError> {
    let slug = slug::transliterated(&raw)?;
    log.set(kind.log_field(), slug);
    Ok(Json(ItemEnvelope {
        data: taxonomy::get_by_poem_count(&state.pg, kind, slug).await?,
    }))
}

#[utoipa::path(
    get,
    path = "/meters",
    tag = "meters",
    operation_id = "meters.list",
    description = "All prosodic meters (al-buhur) with poem and poet counts.",
    responses(
        (status = 200, description = "All meters.", body = ListEnvelope<CountedStats>),
        ListErrors,
    ),
)]
pub(crate) async fn list_meters(
    state: State<AppState>,
    log: Extension<LogHandle>,
) -> Result<Json<ListEnvelope<CountedStats>>, AppError> {
    list_counted_kind(state, log, Counted::Meters).await
}

#[utoipa::path(
    get,
    path = "/meters/{slug}",
    tag = "meters",
    operation_id = "meters.get",
    description = "A single meter with poem and poet counts, by slug.",
    params(
        ("slug" = String, Path, description = "Resource identifier taken from the `slug` field of the matching list endpoint.", pattern = "^[a-z][a-z-]*$", example = "altawil"),
    ),
    responses(
        (status = 200, description = "The requested meter.", body = ItemEnvelope<CountedStats>),
        LookupErrors,
    ),
)]
pub(crate) async fn get_meter(
    state: State<AppState>,
    log: Extension<LogHandle>,
    path: SafePath<String>,
) -> Result<Json<ItemEnvelope<CountedStats>>, AppError> {
    get_counted_kind(state, log, path, Counted::Meters).await
}

#[utoipa::path(
    get,
    path = "/rhymes",
    tag = "rhymes",
    operation_id = "rhymes.list",
    description = "All rhyme letters (al-qawafi) with poem and poet counts.",
    responses(
        (status = 200, description = "All rhymes.", body = ListEnvelope<CountedStats>),
        ListErrors,
    ),
)]
pub(crate) async fn list_rhymes(
    state: State<AppState>,
    log: Extension<LogHandle>,
) -> Result<Json<ListEnvelope<CountedStats>>, AppError> {
    list_counted_kind(state, log, Counted::Rhymes).await
}

#[utoipa::path(
    get,
    path = "/rhymes/{slug}",
    tag = "rhymes",
    operation_id = "rhymes.get",
    description = "A single rhyme with poem and poet counts, by slug.",
    params(
        ("slug" = String, Path, description = "Resource identifier taken from the `slug` field of the matching list endpoint.", pattern = "^[a-z][a-z-]*$", example = "meem"),
    ),
    responses(
        (status = 200, description = "The requested rhyme.", body = ItemEnvelope<CountedStats>),
        LookupErrors,
    ),
)]
pub(crate) async fn get_rhyme(
    state: State<AppState>,
    log: Extension<LogHandle>,
    path: SafePath<String>,
) -> Result<Json<ItemEnvelope<CountedStats>>, AppError> {
    get_counted_kind(state, log, path, Counted::Rhymes).await
}

#[utoipa::path(
    get,
    path = "/eras",
    tag = "eras",
    operation_id = "eras.list",
    description = "All literary eras (al-usur al-adabiyya) with poem and poet counts.",
    responses(
        (status = 200, description = "All eras.", body = ListEnvelope<CountedStats>),
        ListErrors,
    ),
)]
pub(crate) async fn list_eras(
    state: State<AppState>,
    log: Extension<LogHandle>,
) -> Result<Json<ListEnvelope<CountedStats>>, AppError> {
    list_counted_kind(state, log, Counted::Eras).await
}

#[utoipa::path(
    get,
    path = "/eras/{slug}",
    tag = "eras",
    operation_id = "eras.get",
    description = "A single era with poem and poet counts, by slug.",
    params(
        ("slug" = String, Path, description = "Resource identifier taken from the `slug` field of the matching list endpoint.", pattern = "^[a-z][a-z-]*$", example = "abbasi"),
    ),
    responses(
        (status = 200, description = "The requested era.", body = ItemEnvelope<CountedStats>),
        LookupErrors,
    ),
)]
pub(crate) async fn get_era(
    state: State<AppState>,
    log: Extension<LogHandle>,
    path: SafePath<String>,
) -> Result<Json<ItemEnvelope<CountedStats>>, AppError> {
    get_counted_kind(state, log, path, Counted::Eras).await
}

#[utoipa::path(
    get,
    path = "/themes",
    tag = "themes",
    operation_id = "themes.list",
    description = "All thematic categories (al-aghrad) with poem counts.",
    responses(
        (status = 200, description = "All themes.", body = ListEnvelope<PoemCountStats>),
        ListErrors,
    ),
)]
pub(crate) async fn list_themes(
    state: State<AppState>,
    log: Extension<LogHandle>,
) -> Result<Json<ListEnvelope<PoemCountStats>>, AppError> {
    list_poem_counted_kind(state, log, PoemCounted::Themes).await
}

#[utoipa::path(
    get,
    path = "/themes/{slug}",
    tag = "themes",
    operation_id = "themes.get",
    description = "A single theme with poem count, by slug.",
    params(
        ("slug" = String, Path, description = "Resource identifier taken from the `slug` field of the matching list endpoint.", pattern = "^[a-z][a-z-]*$", example = "alnasib"),
    ),
    responses(
        (status = 200, description = "The requested theme.", body = ItemEnvelope<PoemCountStats>),
        LookupErrors,
    ),
)]
pub(crate) async fn get_theme(
    state: State<AppState>,
    log: Extension<LogHandle>,
    path: SafePath<String>,
) -> Result<Json<ItemEnvelope<PoemCountStats>>, AppError> {
    get_poem_counted_kind(state, log, path, PoemCounted::Themes).await
}

#[utoipa::path(
    get,
    path = "/collections",
    tag = "collections",
    operation_id = "collections.list",
    description = "All curated poem collections (al-dawawin) with poem counts.",
    responses(
        (status = 200, description = "All collections.", body = ListEnvelope<PoemCountStats>),
        ListErrors,
    ),
)]
pub(crate) async fn list_collections(
    state: State<AppState>,
    log: Extension<LogHandle>,
) -> Result<Json<ListEnvelope<PoemCountStats>>, AppError> {
    list_poem_counted_kind(state, log, PoemCounted::Collections).await
}

#[utoipa::path(
    get,
    path = "/collections/{slug}",
    tag = "collections",
    operation_id = "collections.get",
    description = "A single collection with poem count, by slug.",
    params(
        ("slug" = String, Path, description = "Resource identifier taken from the `slug` field of the matching list endpoint.", pattern = "^[a-z][a-z-]*$", example = "almuallaqat"),
    ),
    responses(
        (status = 200, description = "The requested collection.", body = ItemEnvelope<PoemCountStats>),
        LookupErrors,
    ),
)]
pub(crate) async fn get_collection(
    state: State<AppState>,
    log: Extension<LogHandle>,
    path: SafePath<String>,
) -> Result<Json<ItemEnvelope<PoemCountStats>>, AppError> {
    get_poem_counted_kind(state, log, path, PoemCounted::Collections).await
}